+++ SugarRecord/Source/CoreData/Reactive/DefaultCoreDataStorage+Reactive.swift
@@ -1,4 +1,3 @@
 import Foundation
-import ReactiveCocoa
 
 extension CoreDataDefaultStorage: ReactiveStorage {}
