+++ SugarRecord/Source/Reactive/ReactiveError.swift
@@ -1,4 +1,3 @@
 import Foundation
 
-
 public enum ReactiveError: ErrorType {}
